+++ Assets/Mochineko/FacialExpressions/Blink/IEyelidMorpher.cs
@@ -2,7 +2,7 @@
 namespace Mochineko.FacialExpressions.Blink
 {
     /// <summary>
-    /// Defines an object that morph eyelid.
+    /// Defines a morpher of eyelid.
     /// </summary>
     public interface IEyelidMorpher
     {
@@ -11,12 +11,14 @@ namespace Mochineko.FacialExpressions.Blink
         /// </summary>
         /// <param name="sample"></param>
         void MorphInto(EyelidSample sample);
+
         /// <summary>
         /// Gets current weight of specified eyelid.
         /// </summary>
         /// <param name="eyelid"></param>
         /// <returns></returns>
         float GetWeightOf(Eyelid eyelid);
+
         /// <summary>
         /// Resets all morphing to default.
         /// </summary>