+++ Assets/Mochineko/FacialExpressions/LipSync/FollowingLipAnimator.cs
@@ -17,17 +17,17 @@ namespace Mochineko.FacialExpressions.LipSync
         private readonly float dt;
         private readonly float initialFollowingVelocity;
         private readonly float followingTime;
+        private readonly Dictionary<Viseme, float> targetWeights = new ();
+        private readonly Dictionary<Viseme, float> followingVelocities = new();
 
         private CancellationTokenSource? animationCanceller;
 
         public FollowingLipAnimator(
             ILipMorpher morpher,
-            int framesPerSecond = 60,
             float initialFollowingVelocity = 0.1f,
             float followingTime = 0.005f)
         {
             this.morpher = morpher;
-            this.dt = 1f / framesPerSecond;
             this.initialFollowingVelocity = initialFollowingVelocity;
             this.followingTime = followingTime;
         }
@@ -39,13 +39,8 @@ namespace Mochineko.FacialExpressions.LipSync
             animationCanceller?.Cancel();
             animationCanceller = CancellationTokenSource
                 .CreateLinkedTokenSource(cancellationToken);
-
-            var targetWeights = new Dictionary<Viseme, float>();
-            var followingVelocities = new Dictionary<Viseme, float>();
-
-            UpdateLoopAsync(targetWeights, followingVelocities, animationCanceller.Token)
-                .Forget();
-
+            
+            targetWeights.Clear();
             morpher.Reset();
 
             foreach (var frame in frames)
@@ -83,37 +78,26 @@ namespace Mochineko.FacialExpressions.LipSync
             animationCanceller = null;
         }
 
-        private async UniTask UpdateLoopAsync(
-            Dictionary<Viseme, float> targetWeights,
-            Dictionary<Viseme, float> velocities,
-            CancellationToken cancellationToken)
+        public void Update()
         {
-            while (!cancellationToken.IsCancellationRequested)
+            foreach (var target in targetWeights)
             {
-                foreach (var target in targetWeights)
-                {
-                    var velocity = velocities[target.Key];
-                    var smoothedWeight = Mathf.Clamp01(Mathf
-                        .SmoothDamp(
-                            current: morpher.GetWeightOf(target.Key),
-                            target: target.Value,
-                            ref velocity,
-                            followingTime));
-                    velocities[target.Key] = velocity;
+                var velocity = followingVelocities[target.Key];
+                var smoothedWeight = Mathf.Clamp01(Mathf
+                    .SmoothDamp(
+                        current: morpher.GetWeightOf(target.Key),
+                        target: target.Value,
+                        ref velocity,
+                        followingTime));
+                followingVelocities[target.Key] = velocity;
 
-                    morpher.MorphInto(new LipSample(target.Key, smoothedWeight));
-                }
-
-                var result = await RelentUniTask.Delay(
-                    TimeSpan.FromSeconds(dt),
-                    cancellationToken: cancellationToken);
-
-                // Cancelled
-                if (result.Failure)
-                {
-                    break;
-                }
+                morpher.MorphInto(new LipSample(target.Key, smoothedWeight));
             }
+        }
+
+        public void Reset()
+        {
+            morpher.Reset();
         }
     }
 }